requirements.txt
streamlit==1.28.0
groq==1.1.2
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.31.0
reportlab==4.0.9
plotly==6.7.0
